benchmark/requirements.txt
openai>=1.0.0
datasets>=2.14.0
tqdm>=4.60.0
Pillow>=10.0.0
PyMuPDF>=1.24.0
